Conflicts:
	.travis.yml
	code/__defines/btime.dm
	code/_helpers/time.dm
	code/_onclick/click.dm
	code/_onclick/hud/alien_larva.dm
	code/_onclick/hud/human.dm
	code/_onclick/hud/other_mobs.dm
	code/_onclick/hud/robot.dm
	code/controllers/emergency_shuttle_controller.dm
	code/controllers/voting.dm
	code/datums/helper_datums/getrev.dm
	code/game/gamemodes/endgame/supermatter_cascade/universe.dm
	code/game/gamemodes/extended/extended.dm
	code/game/gamemodes/gameticker.dm
	code/game/jobs/job/captain.dm
	code/game/jobs/job/engineering.dm
	code/game/jobs/job/medical.dm
	code/game/jobs/job/science.dm
	code/game/jobs/job/security.dm
	code/game/jobs/job/silicon.dm
	code/game/machinery/cryo.dm
	code/game/machinery/nuclear_bomb.dm
	code/game/machinery/telecomms/broadcaster.dm
	code/game/machinery/vending.dm
	code/game/objects/items.dm
	code/game/objects/items/stacks/rods.dm
	code/game/objects/items/weapons/weaponry.dm
	code/game/response_team.dm
	code/game/verbs/who.dm
	code/modules/admin/admin_verbs.dm
	code/modules/admin/verbs/adminpm.dm
	code/modules/client/preference_setup/antagonism/01_candidacy.dm
	code/modules/client/preference_setup/occupation/occupation.dm
	code/modules/clothing/spacesuits/rig/modules/computer.dm
	code/modules/clothing/spacesuits/rig/modules/ninja.dm
	code/modules/clothing/spacesuits/rig/modules/vision.dm
	code/modules/clothing/spacesuits/rig/suits/combat.dm
	code/modules/clothing/spacesuits/rig/suits/ert.dm
	code/modules/clothing/spacesuits/rig/suits/station.dm
	code/modules/clothing/spacesuits/void/station.dm
	code/modules/events/infestation.dm
	code/modules/materials/material_sheets.dm
	code/modules/mob/emote.dm
	code/modules/mob/living/carbon/alien/life.dm
	code/modules/mob/living/carbon/brain/MMI.dm
	code/modules/mob/living/carbon/brain/life.dm
	code/modules/mob/living/carbon/carbon.dm
	code/modules/mob/living/carbon/human/emote.dm
	code/modules/mob/living/carbon/human/human.dm
	code/modules/mob/living/carbon/human/human_defines.dm
	code/modules/mob/living/carbon/human/life.dm
	code/modules/mob/living/carbon/human/species/species.dm
	code/modules/mob/living/carbon/human/stripping.dm
	code/modules/mob/living/carbon/metroid/metroid.dm
	code/modules/mob/living/life.dm
	code/modules/mob/living/silicon/robot/life.dm
	code/modules/mob/living/simple_animal/hostile/mimic.dm
	code/modules/mob/mob.dm
	code/modules/mob/new_player/new_player.dm
	code/modules/mob/observer/freelook/ai/cameranet.dm
	code/modules/mob/observer/freelook/ai/chunk.dm
	code/modules/mob/observer/ghost/ghost.dm
	code/modules/mob/observer/ghost/login.dm
	code/modules/mob/observer/observer.dm
	code/modules/organs/internal/lungs.dm
	code/modules/paperwork/paper.dm
	code/modules/projectiles/gun.dm
	code/modules/projectiles/guns/projectile/automatic.dm
	code/modules/reagents/reagent_containers/food.dm
	code/modules/reagents/reagent_containers/glass/bottle.dm
	code/modules/reagents/reagent_containers/spray.dm
	code/modules/surgery/implant.dm
	code/modules/surgery/robotics.dm
	code/modules/tables/interactions.dm
	code/world.dm
	maps/torch/torch-10.dmm

+++ code/game/objects/items/weapons/weaponry.dm
@@ -12,18 +12,14 @@
 	w_class = 2
 
 /obj/item/weapon/nullrod/attack(mob/M as mob, mob/living/user as mob) //Paste from old-code to decult with a null rod.
-
-	M.attack_log += text("\[[time_stamp()]\] <font color='orange'>Has been attacked with [src.name] by [user.name] ([user.ckey])</font>")
-	user.attack_log += text("\[[time_stamp()]\] <font color='red'>Used the [src.name] to attack [M.name] ([M.ckey])</font>")
-
-	msg_admin_attack("[user.name] ([user.ckey]) attacked [M.name] ([M.ckey]) with [src.name] (INTENT: [uppertext(user.a_intent)]) (<A HREF='?_src_=holder;adminplayerobservecoodjump=1;X=[user.x];Y=[user.y];Z=[user.z]'>JMP</a>)")
+	admin_attack_log(user, M, "Attacked using \a [src]", "Was attacked with \a [src]", "used \a [src] to attack")
 
 	user.setClickCooldown(DEFAULT_ATTACK_COOLDOWN)
 	user.do_attack_animation(M)
 	//if(user != M)
-	if(user.spell_list.len)
-		user.silence_spells(300) //30 seconds
-		user << "<span class='danger'>You've been silenced!</span>"
+	if(M.spell_list.len)
+		M.silence_spells(300) //30 seconds
+		M << "<span class='danger'>You've been silenced!</span>"
 		return
 
 	if (!(istype(user, /mob/living/carbon/human) || ticker) && ticker.mode.name != "monkey")
@@ -67,6 +63,7 @@
 	var/net_type = /obj/effect/energy_net
 
 /obj/item/weapon/energy_net/dropped()
+	..()
 	spawn(10)
 		if(src) qdel(src)
 
@@ -82,10 +79,7 @@
 	var/turf/T = get_turf(M)
 	if(T)
 		var/obj/effect/energy_net/net = new net_type(T)
-		net.layer = M.layer+1
-		M.captured = 1
-		net.affecting = M
-		T.visible_message("[M] was caught in an energy net!")
+		net.capture_mob(M)
 		qdel(src)
 
 	// If we miss or hit an obstacle, we still want to delete the net.
@@ -102,75 +96,34 @@
 	opacity = 0
 	mouse_opacity = 1
 	anchored = 1
+	can_buckle = 0 //no manual buckling or unbuckling
 
 	var/health = 25
-	var/mob/living/affecting = null //Who it is currently affecting, if anyone.
-	var/mob/living/master = null    //Who shot web. Will let this person know if the net was successful.
 	var/countdown = -1
 
 /obj/effect/energy_net/teleport
 	countdown = 60
 
-/obj/effect/energy_net/New()
-	..()
-	processing_objects |= src
+/obj/effect/energy_net/proc/capture_mob(mob/living/M)
+	if(M.buckled)
+		M.buckled.unbuckle_mob()
+	buckle_mob(M)
 
-/obj/effect/energy_net/Destroy()
-
-	if(affecting)
-		var/mob/living/carbon/M = affecting
-		M.anchored = initial(affecting.anchored)
-		M.captured = 0
+/obj/effect/energy_net/post_buckle_mob(mob/living/M)
+	if(buckled_mob)
+		plane = ABOVE_HUMAN_PLANE
+		layer = ABOVE_HUMAN_LAYER
+		visible_message("\The [M] was caught in an energy net!")
+	else
 		M << "You are free of the net!"
-
-	processing_objects -= src
-	..()
+		reset_plane_and_layer()
+		qdel(src)
 
 /obj/effect/energy_net/proc/healthcheck()
-
 	if(health <=0)
 		density = 0
 		src.visible_message("The energy net is torn apart!")
-		qdel(src)
-	return
-
-/obj/effect/energy_net/process()
-
-	if(isnull(affecting) || affecting.loc != loc)
-		qdel(src)
-		return
-
-	// Countdown begin set to -1 will stop the teleporter from firing.
-	// Clientless mobs can be netted but they will not teleport or decrement the timer.
-	var/mob/living/M = affecting
-	if(countdown == -1 || (istype(M) && !M.client))
-		return
-
-	if(countdown > 0)
-		countdown--
-		return
-
-	// TODO: consider removing or altering this; energy nets are useful on their own
-	// merits and the teleportation was never properly implemented; it's halfassed.
-	density = 0
-	invisibility = 101 //Make the net invisible so all the animations can play out.
-	health = INFINITY  //Make the net invincible so that an explosion/something else won't kill it during anims.
-
-	playsound(affecting.loc, 'sound/effects/sparks4.ogg', 50, 1)
-	anim(affecting.loc,affecting,'icons/mob/mob.dmi',,"phaseout",,affecting.dir)
-
-	affecting.visible_message("[affecting] vanishes in a flare of light!")
-
-	if(holdingfacility.len)
-		affecting.loc = pick(holdingfacility)
-
-	affecting << "You appear in a strange place!"
-
-	playsound(affecting.loc, 'sound/effects/phasein.ogg', 25, 1)
-	playsound(affecting.loc, 'sound/effects/sparks2.ogg', 50, 1)
-	anim(affecting.loc,affecting,'icons/mob/mob.dmi',,"phasein",,affecting.dir)
-
-	qdel(src)
+		unbuckle_mob()
 
 /obj/effect/energy_net/bullet_act(var/obj/item/projectile/Proj)
 	health -= Proj.get_structure_damage()
